remove get all bills endpoint, changed all bill endpoints to check if user matches

+++ src/main/java/com/duemate/duemate/service/BillService.java
@@ -10,6 +10,7 @@ import com.duemate.duemate.dto.BillResponse;
 import com.duemate.duemate.dto.CreateBillRequest;
 import com.duemate.duemate.dto.UpdateBillRequest;
 import com.duemate.duemate.exception.BillNotFoundException;
+import com.duemate.duemate.exception.ForbiddenException;
 import com.duemate.duemate.mapper.BillMapper;
 import com.duemate.duemate.model.Bill;
 import com.duemate.duemate.model.BillStatus;
@@ -24,11 +25,11 @@ public class BillService {
 
     private final BillMapper billMapper;
     private final BillRepository billRepository;
-    private final UserService userService;
+    private final CurrentUserService currentUserService;
 
     // POST - Create a bill
     public BillResponse createBill(CreateBillRequest request) {
-        User user = userService.getUserEntityById(request.getUser().getId());
+        User user = currentUserService.getCurrentUser();
 
         BillStatus status = BillStatus.PENDING;
         LocalDate dueDate = request.getDueDate();
@@ -44,29 +45,31 @@ public class BillService {
         return billMapper.convertBillToResponse(bill);
     }
 
-    // GET - Get all bills
-    public List<BillResponse> getAllBills() {
-        List<Bill> bills = billRepository.findAll();
-        return billMapper.convertBillListToResponse(bills);
-    }
-
     // GET - Get a bill by ID
     public BillResponse getBillById(Long id) {
+        User currentUser = currentUserService.getCurrentUser();
         Bill bill = billRepository.findById(id)
                 .orElseThrow(() -> new BillNotFoundException("Bill with id " + id + " not found."));
+        if (!currentUser.getEmail().equals(bill.getUser().getEmail())) {
+            throw new ForbiddenException("You do not have permission to access this resource");
+        }
         return billMapper.convertBillToResponse(bill);
     }
 
     // GET - Get all bills by User
-    public List<BillResponse> getBillsByUser(Long userId) {
-        User user = userService.getUserEntityById(userId);
+    public List<BillResponse> getBillsByUser() {
+        User user = currentUserService.getCurrentUser();
         List<Bill> bills = billRepository.getBillsByUser(user);
         return billMapper.convertBillListToResponse(bills);
     }
 
     // UPDATE - Update a bill
     public BillResponse updateBill(UpdateBillRequest request, Long id) {
+        User user = currentUserService.getCurrentUser();
         Bill bill = getBillEntityById(id);
+        if (!user.getEmail().equals(bill.getUser().getEmail())) {
+            throw new ForbiddenException("You do not have permission to access this resource");
+        }
         bill.setName(request.getName());
         bill.setAmount(request.getAmount());
         bill.setDueDate(request.getDueDate());
@@ -76,7 +79,11 @@ public class BillService {
 
     // UPDATE - Mark a bill as paid
     public BillResponse markBillPaid(Long id) {
+        User user = currentUserService.getCurrentUser();
         Bill bill = getBillEntityById(id);
+        if (!user.getEmail().equals(bill.getUser().getEmail())) {
+            throw new ForbiddenException("You do not have permission to access this resource");
+        }
         bill.setStatus(BillStatus.PAID);
         Bill updatedBill = billRepository.save(bill);
         return billMapper.convertBillToResponse(updatedBill);
@@ -97,7 +104,11 @@ public class BillService {
 
     // DELETE - Delete a bill
     public String deleteBill(Long id) {
+        User user = currentUserService.getCurrentUser();
         Bill bill = getBillEntityById(id);
+        if (!user.getEmail().equals(bill.getUser().getEmail())) {
+            throw new ForbiddenException("You do not have permission to access this resource");
+        }
         billRepository.delete(bill);
         return "Successfully deleted";
     }
